Add files via upload
This task involves editing and running a C program in Ubuntu to calculate the sum of the first N numbers. As shown in image , the program is written in a file named sum.c using the Leafpad editor. It calculates the sum of numbers from 1 to N (set to 5 in this case) using a loop and prints the result. The next steps involve compiling the program with the GCC compiler and executing it in the terminal to display the calculated sum.